Simplify device provider manifest ownership
