+++ packages/device-syncd/README.md
@@ -75,7 +75,7 @@ Strava uses OAuth, polling, and optional app-global webhooks.
 Oura uses OAuth plus refresh tokens and works well in a polling-first mode, so the basic Murph setup does not require Oura webhooks. Once the operator configures the Oura client ID and secret, the end-user flow is just connect once and let scheduled sync keep the account fresh.
 
 The provider lifecycle metadata used here now comes from the shared `@murphai/importers/device-providers/provider-descriptors` surface, so callback paths, default scopes, webhook capabilities, sync windows, metric families, and source-priority hints stay aligned between connector code and snapshot normalization.
-The configured-provider assembly is now derived from one shared provider-manifest registry in `packages/device-syncd/src/config/provider-manifests.ts`, so env specs, config readers, provider factories, provider-owned job definitions, serializable hosted-runtime config, hosted wake hint shaping, and importer wiring all follow the same provider list.
+The configured-provider assembly is derived from one shared, build-safe provider-manifest registry in `packages/device-syncd/src/config/provider-manifests.ts`, so env specs, config readers, provider-owned job definitions, descriptor projection, credential policy, serializable hosted-runtime config, and hosted wake hint shaping follow the same provider facts. Runtime provider factories and importer adapters stay in provider/runtime modules so hosted web can read provider facts without importing the provider implementation graph.
 
 Provider request failures emit a shared, metadata-only diagnostic shape. Logs and hosted runtime apply payloads may include endpoint kind, method, auth placement, body/query field names and counts, upstream status, response shape, and a sanitized provider error code/description. They must not include provider tokens, client secrets, auth codes, raw request bodies, raw response bodies, raw provider paths, query values, or provider account identifiers. New provider transports should use the shared provider diagnostics helpers instead of adding provider-specific ad hoc logging.
